chore: R2 experiment — bug fixes, L4 skills, configs, docs, tests
Source changes (R1 audit fixes + R2 improvements):
- skillq/layers/l1_retrieval/scoring.py
- skillq/runtime/agent.py
- skillq/runtime/container_wiring.py
- skillq/runtime/env_seed.py
- skillq/services/ranking_service.py
- skillq/shared/calls_log.py
- tests/: new tests for calls_log, env_seed, rank, per-trial settings

93 L4-generated skills from R1+R2 experiments

Deleted 30+ old seed skills (replaced by L4 skills)

New configs:
- tb2_skillq_e2e.yaml (3-task validation)
- tb2_skillq_fromscratch_resume.yaml (4-task resume)

Docs:
- doc/tb2_fromscratch_bugs.md (9 bugs documented)
- doc/tb2_fromscratch_report.md (full R1 report)
- doc/experiment_r1_r2_report.md (R1→R2 paper report)

+++ skillq/layers/l1_retrieval/scoring.py
@@ -72,7 +72,14 @@ def cosine(a: Sequence[float], b: Sequence[float]) -> float:
     na = math.sqrt(sum(x * x for x in a)) + 1e-9
     nb = math.sqrt(sum(x * x for x in b)) + 1e-9
     n = min(len(a), len(b))
-    return sum(a[i] * b[i] for i in range(n)) / (na * nb)
+    # Pin return type to Python ``float``. The pure-Python loop
+    # already produces float when ``a`` / ``b`` are lists, but
+    # ``emb_cache`` stores numpy.float32 vectors — without this cast
+    # callers receive ``numpy.float32`` and Pydantic v2's serializer
+    # chokes on it (Bug #1, 2026-06-30:
+    # ``PydanticSerializationError: Unable to serialize unknown type:
+    # <class 'numpy.float32'>`` at /rank → 500 → hook fails open).
+    return float(sum(a[i] * b[i] for i in range(n)) / (na * nb))
 
 
 _cosine = cosine  # legacy private alias used by the inline hook
